reafact : add ball and validate directory and move classes
