+++ src/main/java/baseball/ball/Ball.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.ball;
 
 import java.util.Objects;
