+++ src/main/java/baseball/ball/BallStatus.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.ball;
 
 public enum BallStatus {
     STRIKE, BALL, NOTHING
